add custom_options parameter for mcg workload
