+++ ocs_ci/ocs/fio_artefacts.py
@@ -13,13 +13,20 @@ from ocs_ci.utility.utils import (
 log = logging.getLogger(__name__)
 
 
-def get_mcg_conf(mcg_obj, workload_bucket):
+def get_mcg_conf(mcg_obj, workload_bucket, custom_options=None):
     """
     Basic fio configuration for upgrade utilization for NooBaa S3 bucket.
 
     Args:
         mcg_obj (obj): MCG object, it can be found among fixtures
         workload_bucket (obj): MCG bucket
+        custom_options (dict): Dictionary of lists containing tuples with
+            additional configuration for fio in format:
+            {'section': [('option', 'value'),...],...}
+            e.g.
+            {'global':[('name','bucketname')],'create':[('time_based','1'),('runtime','48h')]}
+            Those values can be added to the config or rewrite already existing
+            values
 
     Returns:
         str: updated fio configuration
@@ -45,6 +52,13 @@ def get_mcg_conf(mcg_obj, workload_bucket):
     config.set('global', 'filename', f"/{workload_bucket[0].name}/object")
     config.set('create', 'time_based', '1')
     config.set('create', 'runtime', '24h')
+
+    # add or overwrite custom values
+    if custom_options:
+        for section in custom_options:
+            for configuration in custom_options[section]:
+                config.set(section, configuration[0], configuration[1])
+
     return config_to_string(config)
